Add fwdlap, reconfigure pydocstyle to be run through flake8 and ignore the new file
