+++ .github/workflows/tests.yml
@@ -55,6 +55,3 @@ jobs:
       run: |
         tox -e black
         tox -e flake8
-
-    - name: Docstring linting
-      run: tox -e pydocstyle
