+++ setup.cfg
@@ -40,7 +40,7 @@ testing =
     pytest-cov==6.0.0
     mypy==1.15.0
     flake8==7.1.1
-    pydocstyle[toml]==6.3.0
+    flake8-docstrings==1.7.0
     tox>=4.24.1
 doc-gen = 
     mkdocs==1.2.3
