+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
 minversion = 4.24.1
-envlist = py312, mypy, black, flake8, pydocstyle
+envlist = py312, mypy, black, flake8
 isolated_build = true
 
 [gh-actions]
 python =
-    3.12: py312, mypy, flake8, pydocstyle
+    3.12: py312, mypy, flake8
 
 [testenv]
 setenv =
@@ -25,13 +25,7 @@ commands =
     black --check --diff --color vmcnet tests
 
 [testenv:flake8]
-deps = flake8
+deps = flake8 flake8-docstrings
 skip_install = true
 commands = 
-    flake8 vmcnet tests
-
-[testenv:pydocstyle]
-deps = pydocstyle[toml]
-skip_install = true
-commands = 
-    pydocstyle vmcnet tests --match='(?!curvature_tags_and_blocks).*\.py'
+    flake8 vmcnet tests --select D --extend-ignore D401
