+++ vmcnet/train/runners.py
@@ -6,7 +6,7 @@ import functools
 import logging
 import os
 import subprocess
-from typing import Any, Optional, Tuple
+from typing import Any, Optional, Tuple, Union
 
 import chex
 import flax
