+++ vmcnet/utils/curvature_tags_and_blocks.py
@@ -1,5 +1,6 @@
 # Types in this file don't pass mypy for a variety of reasons, so don't typecheck it.
 # type: ignore
+# flake8: noqa
 #
 # The rest of this file is are taken directly from DeepMind's FermiNet repository, see
 # https://github.com/deepmind/ferminet/blob/61af370b60179baf01bc1fc5873b8d7f18381600/ferminet/curvature_tags_and_blocks.py
